requirements.txt
flask
flask-cors
chromadb
sentence-transformers
groq
python-dotenv